LayoutTests/platform/mac-sequoia/imported/w3c/web-platform-tests/css/css-overflow/scrollable-overflow-padding-input-expected.txt







PASS .test 1
PASS .test 2
PASS .test 3
FAIL .test 4 assert_equals:
<input type="search" class="test" value="X" data-expected-scroll-height="120">
scrollHeight expected 120 but got 21
PASS .test 5
